Rename tsx file to follow naming convention
